requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0

# QR Code and Barcode scanning
qrcode[pil]==7.4.2
opencv-python==4.9.0.80
pyzbar==0.1.9
pillow==10.0.1

# ISBN lookup and validation
isbnlib==3.10.14
requests==2.31.0

# Core libraries
numpy==1.24.4
cryptography==41.0.7
urllib3==2.0.7

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
